+++ myctrl_movie.cpp
@@ -285,7 +285,6 @@ void film_oversigt_typem::resetallefilm() {
      for(int i=0;i<filmoversigtsize-1;i++) {
        filmoversigt[i].resetfilm();			// back cover
     }
-
 }
 
 
@@ -356,6 +355,8 @@ float film_oversigt_typem::getmovieposition() {
 
 // ****************************************************************************************
 //
+// Next chapter
+//
 // ****************************************************************************************
 
 
@@ -366,6 +367,8 @@ void film_oversigt_typem::next_movie_chapther() {
 
 // ****************************************************************************************
 //
+// Last chapter
+//
 // ****************************************************************************************
 
 void film_oversigt_typem::prevous_movie_chapther() {
@@ -374,6 +377,8 @@ void film_oversigt_typem::prevous_movie_chapther() {
 
 
 // ****************************************************************************************
+//
+// Volume up
 //
 // ****************************************************************************************
 
@@ -385,6 +390,8 @@ void film_oversigt_typem::volumeup() {
 
 // ****************************************************************************************
 //
+// Volume down
+//
 // ****************************************************************************************
 
 void film_oversigt_typem::volumedown() {
@@ -394,6 +401,8 @@ void film_oversigt_typem::volumedown() {
 
 
 // ****************************************************************************************
+//
+// Set Volume
 //
 // ****************************************************************************************
