+++ myctrl_movie.h
@@ -112,11 +112,12 @@ class film_oversigt_typem : vlc_controller {
 //      libvlc_instance_t *vlc_inst;              //
 //      libvlc_media_player_t *vlc_mp;            //
       float getmovieposition();
-      bool show_search_view;
+      bool show_search_view;                                                    // Do we search for title
+      bool film_is_playing;                                                     // playing ?
   public:
+      bool is_film_playing() { return (film_is_playing); }
       bool get_search_view() { return (show_search_view); }
       void set_search_view(bool val) { show_search_view=val; }
-      bool film_is_playing;                                                     // playing ?
       film_oversigt_type *filmoversigt;                                         // all movie array to show in movie overview
       void sortfilm(int type);
       unsigned int film_antal(void) { return(filmoversigt_antal); }
