+++ utility.h
@@ -7,7 +7,7 @@ int clearscreen();
 int getuserhomedir(char *homedir);
 int check_zerro_bytes_file(char *filename);
 unsigned int hourtounixtime(int hour);
-
 void write_logfile(char *logentry);
-
+char *b64_encode(const unsigned char *in, size_t len);
+size_t b64_encoded_size(size_t inlen);
 #endif
